update ffind, sigaction

+++ roadprj/types.h
@@ -1,9 +1,9 @@
 #ifndef RTYPES_H
 #define RTYPES_H
 #include <stdint.h>
-typedef uint8_t car_id;
-typedef uint8_t trigger_event;
-typedef uint8_t car_status;
+typedef uint32_t car_id;
+typedef uint32_t trigger_event;
+typedef uint32_t car_status;
 typedef uint64_t gps_time;
 typedef uint64_t gps_x;
 typedef uint64_t gps_y;
@@ -11,9 +11,9 @@ typedef struct {
     gps_x x;
     gps_y y;
 } gps_coord;
-typedef uint8_t gps_speed;
-typedef uint8_t gps_direction;
-typedef uint8_t gps_valid;
+typedef uint32_t gps_speed;
+typedef uint32_t gps_direction;
+typedef uint32_t gps_valid;
 typedef struct {
     gps_time time;
     gps_coord coord;
